// src/app/models/enums/route/route-title.enum.ts
/** Заголовки пунктов меню */
export enum RouteTitleEnum {
  main = 'Главная',
  angular = 'Angular',
  forms = 'Формы',
  javascript = 'JavaScript',
  rxjs = 'RxJS',
  rxjsHoo = 'RxJS-HOO',
  typescript = 'TypeScript',
  tsUtils = 'TS Utils',
  tsGen = 'TS Generic',
  zone = 'Zone',
}